added new ssl version

+++ SQL/install-sql.ps1
@@ -161,15 +161,16 @@ Switch ($SQLVER)
 $Isopath = Join-path $SQL_BASEDir $Iso_File
 Write-Verbose $Isopath
 .$Nodescriptdir\test-setup -setup $SQL_BASEVER -setuppath $Isopath
-Write-Host -ForegroundColor Gray "Copying $SQL_BASEVER ISO locally"
-Copy-Item $Isopath -Destination "$env:USERPROFILE\Downloads"
+if (!(Test-Path "$env:USERPROFILE\Downloads\$Iso_File"))
+	{
+	Write-Host -ForegroundColor Gray "Copying $SQL_BASEVER ISO locally"
+	Copy-Item $Isopath -Destination "$env:USERPROFILE\Downloads"
+	}
 $Temp_Iso = "$env:USERPROFILE\Downloads\$Iso_File"
 $ismount = Mount-DiskImage -ImagePath $Temp_Iso -PassThru
 $Driveletter = (Get-Volume | where { $_.size -eq $ismount.Size}).driveletter
 $Setupcmd = "setup.exe"        
 $Setuppath = "$($Driveletter):\$Setupcmd"
-
-
 if (!$DefaultDBpath.IsPresent)
     {
     $Diskparameter = "/SQLUSERDBDIR=m:\ /SQLUSERDBLOGDIR=n:\ /SQLTEMPDBDIR=o:\ /SQLTEMPDBLOGDIR=p:\"
